requirements.txt
tensorflow
keras
opencv-python
pillow
pycryptodome
numpy
scikit-learn
